+++ protobuf/identitypb/identity.go
@@ -21,6 +21,37 @@ func New(name string, key *uuidpb.UUID) *Identity {
 	return x
 }
 
+// Parse returns a new identity with the given name and key by parsing the key
+// as a UUID.
+func Parse(name, key string) (*Identity, error) {
+	k, err := uuidpb.Parse(key)
+	if err != nil {
+		return nil, fmt.Errorf("invalid key: %w", err)
+	}
+
+	x := &Identity{
+		Name: name,
+		Key:  k,
+	}
+
+	if err := x.Validate(); err != nil {
+		return nil, err
+	}
+
+	return x, nil
+}
+
+// MustParse returns a new identity with the given name and key by parsing the
+// key as a UUID, or panics if unable to do so.
+func MustParse(name, key string) *Identity {
+	x, err := Parse(name, key)
+	if err != nil {
+		panic(err)
+	}
+
+	return x
+}
+
 // Validate returns an error if x is invalid.
 //
 // It does not perform UTF-8 validation on the name. This should be validated by
@@ -51,5 +82,5 @@ func (x *Identity) Format(f fmt.State, verb rune) {
 
 // Equal returns true if x and id are equal.
 func (x *Identity) Equal(id *Identity) bool {
-	return x.GetName() == id.GetName() && x.GetKey() == id.GetKey()
+	return x.GetName() == id.GetName() && x.GetKey().Equal(id.GetKey())
 }