+++ CHANGELOG.md
@@ -11,6 +11,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
+## [0.19.4] - 2025-10-11
+
+### Added
+
+- Added `identitypb.Parse()` and `MustParse()`.
+
+### Fixed
+
+- Fixed `identitypb.Identity.Equal()` to correctly compare identity keys.
+
 ## [0.19.3] - 2025-10-11
 
 ### Added
@@ -301,6 +311,7 @@ which itself includes a large numbers of breaking changes.
 [0.19.1]: https://github.com/dogmatiq/enginekit/releases/v0.19.1
 [0.19.2]: https://github.com/dogmatiq/enginekit/releases/v0.19.2
 [0.19.3]: https://github.com/dogmatiq/enginekit/releases/v0.19.3
+[0.19.4]: https://github.com/dogmatiq/enginekit/releases/v0.19.4
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD